+++ jni/src/simd/similarity_function/similarity_function.cpp
@@ -1,0 +1,14 @@
+/*
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * The OpenSearch Contributors require contributions made to
+ * this file be licensed under the Apache-2.0 license or a
+ * compatible open source license.
+ *
+ * Modifications Copyright OpenSearch Contributors. See
+ * GitHub history for details.
+ */
+
+// Today, we're sitting on V1 implementation.
+// For V2, refer to https://github.com/opensearch-project/k-NN/issues/2875 for more details.
+#include "default_simd_similarity_function.cpp"