+++ src/test/java/org/opensearch/knn/memoryoptsearch/MMapByteVectorValuesTests.java
@@ -53,7 +53,11 @@ public class MMapByteVectorValuesTests extends LuceneTestCase {
 
             // Read validation
             try (final IndexInput input = directory.openInput(fileName, IOContext.DEFAULT)) {
-                final long[] addressAndSize = MemorySegmentAddressExtractorUtil.tryExtractAddressAndSize(input);
+                final long[] addressAndSize = MemorySegmentAddressExtractorUtil.tryExtractAddressAndSize(
+                    input,
+                    0,
+                    numVectors * dimension * Float.BYTES
+                );
                 assertNotNull(addressAndSize);
                 final MMapByteVectorValues values = new MMapByteVectorValues(
                     input,
@@ -65,7 +69,6 @@ public class MMapByteVectorValuesTests extends LuceneTestCase {
                 );
 
                 // Ensure properties are correct.
-                assertEquals(oneVectorByteSize, values.getOneVectorByteSize());
                 assertEquals(oneVectorByteSize, values.getVectorByteLength());
                 assertEquals(addressAndSize, values.getAddressAndSize());